+++ packages/router/router.ts
@@ -3,7 +3,8 @@ import {
   Action as HistoryAction,
   createLocation,
   createPath,
-  createURL,
+  createClientSideURL,
+  invariant,
   parsePath,
 } from "./history";
 import type {
@@ -28,7 +29,6 @@ import {
   ResultType,
   convertRoutesToDataRoutes,
   getPathContributingMatches,
-  invariant,
   isRouteErrorResponse,
   joinPaths,
   matchRoutes,
@@ -316,8 +316,14 @@ export interface StaticHandlerContext {
  */
 export interface StaticHandler {
   dataRoutes: AgnosticDataRouteObject[];
-  query(request: Request): Promise<StaticHandlerContext | Response>;
-  queryRoute(request: Request, routeId?: string): Promise<any>;
+  query(
+    request: Request,
+    opts?: { requestContext?: unknown }
+  ): Promise<StaticHandlerContext | Response>;
+  queryRoute(
+    request: Request,
+    opts?: { routeId?: string; requestContext?: unknown }
+  ): Promise<any>;
 }
 
 /**
@@ -913,7 +919,7 @@ export function createRouter(init: RouterInit): Router {
 
     // Create a controller/Request for this navigation
     pendingNavigationController = new AbortController();
-    let request = createRequest(
+    let request = createClientSideRequest(
       location,
       pendingNavigationController.signal,
       opts && opts.submission
@@ -952,6 +958,9 @@ export function createRouter(init: RouterInit): Router {
         ...opts.submission,
       };
       loadingNavigation = navigation;
+
+      // Create a GET request for the loaders
+      request = new Request(request.url, { signal: request.signal });
     }
 
     // Call loaders
@@ -1296,7 +1305,11 @@ export function createRouter(init: RouterInit): Router {
 
     // Call the action for the fetcher
     let abortController = new AbortController();
-    let fetchRequest = createRequest(path, abortController.signal, submission);
+    let fetchRequest = createClientSideRequest(
+      path,
+      abortController.signal,
+      submission
+    );
     fetchControllers.set(key, abortController);
 
     let actionResult = await callLoaderOrAction(
@@ -1343,7 +1356,7 @@ export function createRouter(init: RouterInit): Router {
     // Start the data load for current matches, or the next location if we're
     // in the middle of a navigation
     let nextLocation = state.navigation.location || state.location;
-    let revalidationRequest = createRequest(
+    let revalidationRequest = createClientSideRequest(
       nextLocation,
       abortController.signal
     );
@@ -1498,7 +1511,7 @@ export function createRouter(init: RouterInit): Router {
 
     // Call the loader for this fetcher route match
     let abortController = new AbortController();
-    let fetchRequest = createRequest(path, abortController.signal);
+    let fetchRequest = createClientSideRequest(path, abortController.signal);
     fetchControllers.set(key, abortController);
     let result: DataResult = await callLoaderOrAction(
       "loader",
@@ -1599,17 +1612,17 @@ export function createRouter(init: RouterInit): Router {
       "Expected a location on the redirect navigation"
     );
 
-    if (
-      redirect.external &&
-      typeof window !== "undefined" &&
-      typeof window.location !== "undefined"
-    ) {
-      if (replace) {
-        window.location.replace(redirect.location);
-      } else {
-        window.location.assign(redirect.location);
+    // Check if this an external redirect that goes to a new origin
+    if (typeof window?.location !== "undefined") {
+      let newOrigin = createClientSideURL(redirect.location).origin;
+      if (window.location.origin !== newOrigin) {
+        if (replace) {
+          window.location.replace(redirect.location);
+        } else {
+          window.location.assign(redirect.location);
+        }
+        return;
       }
-      return;
     }
 
     // There's no need to abort on redirects, since we don't detect the
@@ -1672,7 +1685,7 @@ export function createRouter(init: RouterInit): Router {
       ...fetchersToLoad.map(([, href, match, fetchMatches]) =>
         callLoaderOrAction(
           "loader",
-          createRequest(href, request.signal),
+          createClientSideRequest(href, request.signal),
           match,
           fetchMatches,
           router.basename
@@ -1936,7 +1949,8 @@ export function unstable_createStaticHandler(
    * return it directly.
    */
   async function query(
-    request: Request
+    request: Request,
+    { requestContext }: { requestContext?: unknown } = {}
   ): Promise<StaticHandlerContext | Response> {
     let url = new URL(request.url);
     let method = request.method.toLowerCase();
@@ -1980,8 +1994,8 @@ export function unstable_createStaticHandler(
       };
     }
 
-    let result = await queryImpl(request, location, matches);
-    if (result instanceof Response) {
+    let result = await queryImpl(request, location, matches, requestContext);
+    if (isResponse(result)) {
       return result;
     }
 
@@ -2011,7 +2025,13 @@ export function unstable_createStaticHandler(
    * code.  Examples here are 404 and 405 errors that occur prior to reaching
    * any user-defined loaders.
    */
-  async function queryRoute(request: Request, routeId?: string): Promise<any> {
+  async function queryRoute(
+    request: Request,
+    {
+      routeId,
+      requestContext,
+    }: { requestContext?: unknown; routeId?: string } = {}
+  ): Promise<any> {
     let url = new URL(request.url);
     let method = request.method.toLowerCase();
     let location = createLocation("", createPath(url), null, "default");
@@ -2038,8 +2058,14 @@ export function unstable_createStaticHandler(
       throw getInternalRouterError(404, { pathname: location.pathname });
     }
 
-    let result = await queryImpl(request, location, matches, match);
-    if (result instanceof Response) {
+    let result = await queryImpl(
+      request,
+      location,
+      matches,
+      requestContext,
+      match
+    );
+    if (isResponse(result)) {
       return result;
     }
 
@@ -2061,6 +2087,7 @@ export function unstable_createStaticHandler(
     request: Request,
     location: Location,
     matches: AgnosticDataRouteMatch[],
+    requestContext: unknown,
     routeMatch?: AgnosticDataRouteMatch
   ): Promise<Omit<StaticHandlerContext, "location" | "basename"> | Response> {
     invariant(
@@ -2074,13 +2101,19 @@ export function unstable_createStaticHandler(
           request,
           matches,
           routeMatch || getTargetMatch(matches, location),
+          requestContext,
           routeMatch != null
         );
         return result;
       }
 
-      let result = await loadRouteData(request, matches, routeMatch);
-      return result instanceof Response
+      let result = await loadRouteData(
+        request,
+        matches,
+        requestContext,
+        routeMatch
+      );
+      return isResponse(result)
         ? result
         : {
             ...result,
@@ -2110,6 +2143,7 @@ export function unstable_createStaticHandler(
     request: Request,
     matches: AgnosticDataRouteMatch[],
     actionMatch: AgnosticDataRouteMatch,
+    requestContext: unknown,
     isRouteRequest: boolean
   ): Promise<Omit<StaticHandlerContext, "location" | "basename"> | Response> {
     let result: DataResult;
@@ -2117,7 +2151,7 @@ export function unstable_createStaticHandler(
     if (!actionMatch.route.action) {
       let error = getInternalRouterError(405, {
         method: request.method,
-        pathname: createURL(request.url).pathname,
+        pathname: new URL(request.url).pathname,
         routeId: actionMatch.route.id,
       });
       if (isRouteRequest) {
@@ -2135,7 +2169,8 @@ export function unstable_createStaticHandler(
         matches,
         basename,
         true,
-        isRouteRequest
+        isRouteRequest,
+        requestContext
       );
 
       if (request.signal.aborted) {
@@ -2185,9 +2220,15 @@ export function unstable_createStaticHandler(
       // Store off the pending error - we use it to determine which loaders
       // to call and will commit it when we complete the navigation
       let boundaryMatch = findNearestBoundary(matches, actionMatch.route.id);
-      let context = await loadRouteData(request, matches, undefined, {
-        [boundaryMatch.route.id]: result.error,
-      });
+      let context = await loadRouteData(
+        request,
+        matches,
+        requestContext,
+        undefined,
+        {
+          [boundaryMatch.route.id]: result.error,
+        }
+      );
 
       // action status codes take precedence over loader status codes
       return {
@@ -2202,7 +2243,9 @@ export function unstable_createStaticHandler(
       };
     }
 
-    let context = await loadRouteData(request, matches);
+    // Create a GET request for the loaders
+    let loaderRequest = new Request(request.url, { signal: request.signal });
+    let context = await loadRouteData(loaderRequest, matches, requestContext);
 
     return {
       ...context,
@@ -2220,6 +2263,7 @@ export function unstable_createStaticHandler(
   async function loadRouteData(
     request: Request,
     matches: AgnosticDataRouteMatch[],
+    requestContext: unknown,
     routeMatch?: AgnosticDataRouteMatch,
     pendingActionError?: RouteData
   ): Promise<
@@ -2235,7 +2279,7 @@ export function unstable_createStaticHandler(
     if (isRouteRequest && !routeMatch?.route.loader) {
       throw getInternalRouterError(400, {
         method: request.method,
-        pathname: createURL(request.url).pathname,
+        pathname: new URL(request.url).pathname,
         routeId: routeMatch?.route.id,
       });
     }
@@ -2268,7 +2312,8 @@ export function unstable_createStaticHandler(
           matches,
           basename,
           true,
-          isRouteRequest
+          isRouteRequest,
+          requestContext
         )
       ),
     ]);
@@ -2526,9 +2571,9 @@ function shouldRevalidateLoader(
   isRevalidationRequired: boolean,
   actionResult: DataResult | undefined
 ) {
-  let currentUrl = createURL(currentLocation);
+  let currentUrl = createClientSideURL(currentLocation);
   let currentParams = currentMatch.params;
-  let nextUrl = createURL(location);
+  let nextUrl = createClientSideURL(location);
   let nextParams = match.params;
 
   // This is the default implementation as to when we revalidate.  If the route
@@ -2571,7 +2616,8 @@ async function callLoaderOrAction(
   matches: AgnosticDataRouteMatch[],
   basename = "/",
   isStaticRequest: boolean = false,
-  isRouteRequest: boolean = false
+  isRouteRequest: boolean = false,
+  requestContext?: unknown
 ): Promise<DataResult> {
   let resultType;
   let result;
@@ -2590,7 +2636,7 @@ async function callLoaderOrAction(
     );
 
     result = await Promise.race([
-      handler({ request, params: match.params }),
+      handler({ request, params: match.params, context: requestContext }),
       abortPromise,
     ]);
 
@@ -2607,7 +2653,7 @@ async function callLoaderOrAction(
     request.signal.removeEventListener("abort", onReject);
   }
 
-  if (result instanceof Response) {
+  if (isResponse(result)) {
     let status = result.status;
 
     // Process redirects
@@ -2618,17 +2664,20 @@ async function callLoaderOrAction(
         "Redirects returned/thrown from loaders/actions must have a Location header"
       );
 
-      // Check if this an external redirect that goes to a new origin
-      let external = createURL(location).origin !== createURL("/").origin;
+      let isAbsolute =
+        /^[a-z+]+:\/\//i.test(location) || location.startsWith("//");
 
       // Support relative routing in internal redirects
-      if (!external) {
+      if (!isAbsolute) {
         let activeMatches = matches.slice(0, matches.indexOf(match) + 1);
         let routePathnames = getPathContributingMatches(activeMatches).map(
           (match) => match.pathnameBase
         );
-        let requestPath = createURL(request.url).pathname;
-        let resolvedLocation = resolveTo(location, routePathnames, requestPath);
+        let resolvedLocation = resolveTo(
+          location,
+          routePathnames,
+          new URL(request.url).pathname
+        );
         invariant(
           createPath(resolvedLocation),
           `Unable to resolve redirect location: ${location}`
@@ -2658,7 +2707,6 @@ async function callLoaderOrAction(
         status,
         location,
         revalidate: result.headers.get("X-Remix-Revalidate") !== null,
-        external,
       };
     }
 
@@ -2708,12 +2756,15 @@ async function callLoaderOrAction(
   return { type: ResultType.data, data: result };
 }
 
-function createRequest(
+// Utility method for creating the Request instances for loaders/actions during
+// client-side navigations and fetches.  During SSR we will always have a
+// Request instance from the static handler (query/queryRoute)
+function createClientSideRequest(
   location: string | Location,
   signal: AbortSignal,
   submission?: Submission
 ): Request {
-  let url = createURL(stripHashFromPath(location)).toString();
+  let url = createClientSideURL(stripHashFromPath(location)).toString();
   let init: RequestInit = { signal };
 
   if (submission) {
@@ -2954,12 +3005,10 @@ function getInternalRouterError(
     pathname,
     routeId,
     method,
-    message,
   }: {
     pathname?: string;
     routeId?: string;
     method?: string;
-    message?: string;
   } = {}
 ) {
   let statusText = "Unknown Server Error";
@@ -3034,8 +3083,18 @@ function isRedirectResult(result?: DataResult): result is RedirectResult {
   return (result && result.type) === ResultType.redirect;
 }
 
+function isResponse(value: any): value is Response {
+  return (
+    value != null &&
+    typeof value.status === "number" &&
+    typeof value.statusText === "string" &&
+    typeof value.headers === "object" &&
+    typeof value.body !== "undefined"
+  );
+}
+
 function isRedirectResponse(result: any): result is Response {
-  if (!(result instanceof Response)) {
+  if (!isResponse(result)) {
     return false;
   }
 
@@ -3047,7 +3106,7 @@ function isRedirectResponse(result: any): result is Response {
 function isQueryRouteResponse(obj: any): obj is QueryRouteResponse {
   return (
     obj &&
-    obj.response instanceof Response &&
+    isResponse(obj.response) &&
     (obj.type === ResultType.data || ResultType.error)
   );
 }